requirement.txt
fastapi
uvicorn
python-jose
passlib[bcrypt]
pydantic
email-validator
bcrypt==4.3.0
python-dotenv